Camera location now double rather than float. Fixed login location specification

+++ src/LookingGlass.Renderer.Ogre/Ogr.cs
@@ -145,7 +145,7 @@ static class Ogr {
     // =============================================================================
     [DllImport("LookingGlassOgre", CallingConvention = CallingConvention.Cdecl)]
     public static extern void UpdateCamera(
-        float px, float py, float pz, 
+        double px, double py, double pz, 
         float dw, float dx, float dy, float dz,
         float nearClip, float farClip, float aspect);
     [DllImport("LookingGlassOgre", CallingConvention = CallingConvention.Cdecl)]
